The per-keyframe timestamp array becomes a Float64Array, removing the heap
BigInt that every fractional-index Hermite read used to box (Number(ts[i+1] -
ts[i]) ran inside useFrame for every body in every consumer, so playback was
churning on the order of 50 to 200 BigInt allocations per frame on a path the
render-loop rules require allocation-free). Unix-ms epochs are exact in float64
far below the 2^53 ceiling, so the decode and Hermite outputs are bit-identical
and the wire still reads its int64 start with getBigInt64 unchanged: no
backend, DTO, OpenAPI, or wire-version change. Four test fixtures beyond the
originally scoped set also carried BigInt64Array timestamps and were migrated.
A new guard test pins the representation and the year-9999 exactness bound.
Gates: tsc, eslint, 358 vitest tests, and next build all green.

L8 was the last open audit finding, so every one of the review findings is now
resolved (or, for the ones closed as deliberate documentation corrections,
accounted for). The audit is closed out.

+++ frontend/src/app/store/middleware/simulationRequestThunk.ts
@@ -18,7 +18,7 @@ interface ChunkPayload {
   bodyCount: number;
   timestepCount: number;
   positions: Float64Array;
-  timestamps: BigInt64Array;
+  timestamps: Float64Array;
   mu: Record<string, number>;
   deltaERelative: Float32Array;
   dp853AvgStepSeconds: number | null;
